LayoutTests/svg/dom/altGlyph-dom-expected.txt

一&M000076; 丁
This test checks that SVG altGlyph elements create the appropriate DOM object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS altGlyph.tagName is "altGlyph"
PASS altGlyph.namespaceURI is "http://www.w3.org/2000/svg"
PASS successfullyParsed is true

TEST COMPLETE
